# WIP: use WASM for hashing