--- a/src/index.html
+++ b/src/index.html
@@ -358,8 +358,6 @@
         <script src="https://cdnjs.cloudflare.com/ajax/libs/humanize-duration/3.14.0/humanize-duration.min.js" integrity="sha256-lfP2nyh9CrETTh3qI+BT/1bJCHmYxyMXQ4QadQ5Befo="
             crossorigin="anonymous"></script>
 
-        <script src="https://cdnjs.cloudflare.com/ajax/libs/asmCrypto/2.0.1/asmcrypto.all.es5.min.js" crossorigin="anonymous"></script>
-
         <script type="module" src="js/main.js"></script>
     </body>