This is faster in every browser but there’s still a fair
amount of variation depending on the browser and we 
presumably need to spend some time troubleshooting 
performance:

* Safari goes from 74MB/s to 98MB/s on large files
* Firefox goes from 6MB/s to ~11MB/s on the same files
* Chrome goes from 59MB/s to 93MB/s on the same files

(sha256deep is ~131MB/s on the same files)

Major questions before we can consider merging this:

* How comfortable are we with the Rust / wasm toolchain? This seems very appealing because it allows us to support new algorithms with reasonable performance rather than shopping around for JavaScript versions.
* What's the best way to package the WASM hashing code so we don't need to include a build artifact in this repo?